Move all the listing utilities under utils/listing.js. Also overwrite listing data
